Add _getLogsAsync on contract_wrapper
